requirements.txt
# Data collection
requests==2.31.0
pandas==2.2.0  # Changed from 2.1.4
numpy==1.26.4  # Changed from 1.26.2

# F1 specific
fastf1==3.3.9  # Changed from 3.3.6

# Environment management
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# Progress bars
tqdm==4.66.1

# Data storage
pyarrow==14.0.2  # Changed from 14.0.1